+++ tests/Fixtures/DocBlockType/Phpstan/PhpstanArrayCollectionShape.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Collection;
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Phpstan;
 
 /**
- * @phpstan-type Settings array<int, Product>
+ * @phpstan-type Settings array<int, ProductType>
  */
 final class PhpstanArrayCollectionShape
 {
