+++ tests/Fixtures/DocBlockType/Phpstan/ProductType.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JMS\Serializer\Tests\Fixtures\DocBlockType\Phpstan;
+
+class ProductType
+{
+}
